SUMMARY.md: remove almost all structure
It seems that GitBook v2 doens’t properly handle links to anchors within pages
(e.g. `foo.md#section`) in this file anymore. It transforms the anchor
fragments into sub-paths (`foo/section`), which then appear as either TOC
nodes or empty in the rendered doc. So basically it's all confusing and broken
if you use anchors. Too bad, since this TOC was spelled out in detail!

One potential solution would be to break the current docs into lots of smaller
files, since those kinds of links seem to work OK. Or, maybe there's a way to
get it to deal with anchors correctly.