examples.md: rename from samples.md and update
At the moment, if we want our examples to work live, we have to host them on a
webserver that doesn't insist on upgrading to HTTPS. Which means, annoyingly,
that they have to live in another repo. I've set that all up, and now the
listing of example code mainly just points to that separate domain.